'Form' and 'canvas' class set up
Add controls to the Windows Form. These are:
- Program Input Area (TextBox)
- Command Line (TextBox)
- Drawing Area (PictureBox)
- Run Button
Disable Run Button if Program TextBox is empty. This is
to circumvent validation checks for programs.

Add several attributes to the 'Canvas' class.
Set up properties for this class to enable
other classes to get and set the attributes of the canvas
in the future.