Refactor code in classes derived from 'Command'
Change parameter input from user to be additionally
split based on a comma. In the current system, this
means splitting the user input by space (" ") and then
by comma (",").

Consider creating a separate class purely for parsing
parameters.

Refactor code in 'Colour' class to use a switch
statement instead of an if / else-if statement.