Implement command and parameter error reporting
Change some constructors of classes to be consistent
with using setters. Add unique variables, relevant to
specific class. This will result in no list of parameter,
but can be resolved later. Change parameter names in
'Operation' method to be consistent with variable additions.

Modify 'ParseParameters' methods
to add to list collection of errors, rather than throwing
an exception.

Check for errors in command's 'Operation' method before
execution of drawing - a message box of the error will
return if error list collection is not empty.

Add constants for error messages in 'Command' class. Remove
obsolete variables in the same class.

Return message box in command factory instead of throwing
exception, if command does not exist.

Remove 'action.Operation' in form class as it was drawing
twice.

TODO: Current error reporting system does not account for
multiline program. Extract error reporting to a different class
and create methods for duplicate code.