Simplify console error reporting in program
Change constructor of 'ConsoleDisplayError' class to
only have one parameter. This will simplify error reporting
for the time being.

TODO: Consider having an instance of the class 'ConsoleDisplayError'
in which you pass the error to, and then it executes all of the
errors. This may be useful for reporting errors for the user
program.