Implement 'User Program' class and code refactor
The 'User Program' class decouples the code from the 'Run'
class that executes drawing operations of a user input program
from the program textbox. This is done by storing commands as an
object in an arraylist. This also allows the ability to assign
line numbers to error messages, based on the indexes in the arraylist.

As a result of these changes, error reporting code in other classes had to
be refactored, to distinguish between command line error reporting, and
program error reporting.