Add command to syntax check before program is run
Add a button and command that the user can use to check
the syntax of the program before execution.

Also, due to recent refactoring of 'Run' and 'UserProgram' classes
for easier unit testing, resolved bug in which run and syntax command
weren't working through the command line.

TODO: Implement conditional statements that the user can use in their
program.