Add loop command feature to application
Add ability for user to loop commands in user program.
However, due to the implementation method, this will result
in the syntax report not outputting the correct program lines.

If you want syntax error reporting to work correctly, you will
need to refactor the whole system so that you use program counters
and flags, so that the program lines are not modified.